requirements.txt
# aws-cdk-lib==2.10.0
# constructs>=10.0.0,<11.0.0
# npm
# node
# cdk
# aws_cdk.ec2
# aws_cdk.core
aws-cdk.core
aws-cdk.aws-ec2